requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
qdrant-client>=1.7.0
openai>=1.12.0
tqdm>=4.66.0
python-dotenv>=1.0.0 